[Hamza/Rudra] Added WebCrypto encryption for syncQueue data
